refactror: move about and 404 to the pages folder
