+++ src/components/index.tsx
@@ -4,8 +4,8 @@ export { Loader } from './loader';
 export { Search } from './search';
 export { Description } from './description';
 export { Home } from './home';
-export { About } from './about';
+export { About } from '../pages/about';
 export { Header } from './header';
-export { NotFound } from './404';
+export { NotFound } from '../pages/404';
 export { Selected } from './selected';
 export { Flyout } from './flyout';
